refactor(leaderboards): extract leaderboard persistence service
