+++ api/services/leaderboard/leaderboard.ts
@@ -1,0 +1,309 @@
+import type { MatchStats } from "halo-infinite-api";
+import { Preconditions } from "@guilty-spark/shared/base/preconditions";
+import { getDurationInSeconds } from "@guilty-spark/shared/halo/duration";
+import { getSafeRatioValue } from "@guilty-spark/shared/halo/stat-formatting";
+import { getPlayerXuid } from "@guilty-spark/shared/halo/match-stats";
+import type { DatabaseService } from "../database/database";
+import type { LeaderboardSeriesRow } from "../database/types/leaderboard_series";
+import type { LeaderboardSeriesPlayersRow } from "../database/types/leaderboard_series_players";
+import type { LeaderboardGamesRow } from "../database/types/leaderboard_games";
+import type { LeaderboardGamePlayersRow } from "../database/types/leaderboard_game_players";
+import type { NeatQueueConfigRow } from "../database/types/neat_queue_config";
+import type { HaloService } from "../halo/halo";
+import type { LogService } from "../log/types";
+import type { NeatQueueMatchCompletedRequest } from "../neatqueue/types";
+
+export interface LeaderboardServiceOpts {
+  databaseService: DatabaseService;
+  haloService: HaloService;
+  logService: LogService;
+}
+
+export class LeaderboardService {
+  private readonly databaseService: DatabaseService;
+  private readonly haloService: HaloService;
+  private readonly logService: LogService;
+
+  constructor({ databaseService, haloService, logService }: LeaderboardServiceOpts) {
+    this.databaseService = databaseService;
+    this.haloService = haloService;
+    this.logService = logService;
+  }
+
+  async persistSeriesData({
+    request,
+    neatQueueConfig,
+    series,
+    locale,
+  }: {
+    request: NeatQueueMatchCompletedRequest;
+    neatQueueConfig: NeatQueueConfigRow;
+    series: MatchStats[];
+    locale: string;
+  }): Promise<void> {
+    if (series.length === 0 || request.winning_team_index === -1) {
+      return;
+    }
+
+    try {
+      const sortedSeries = [...series].sort((left, right) =>
+        left.MatchInfo.StartTime.localeCompare(right.MatchInfo.StartTime),
+      );
+      const firstMatch = Preconditions.checkExists(sortedSeries[0]);
+      const lastMatch = Preconditions.checkExists(sortedSeries[sortedSeries.length - 1]);
+      const nowEpoch = Math.floor(Date.now() / 1000);
+      const startedAt = this.toEpochSeconds(firstMatch.MatchInfo.StartTime) ?? nowEpoch;
+      const completedAt = this.toEpochSeconds(lastMatch.MatchInfo.EndTime) ?? nowEpoch;
+
+      const seriesRow: LeaderboardSeriesRow = {
+        GuildId: request.guild,
+        QueueNumber: request.match_number,
+        QueueChannelId: neatQueueConfig.ChannelId,
+        ResultsChannelId: neatQueueConfig.ResultsChannelId,
+        StartedAt: startedAt,
+        CompletedAt: completedAt,
+        WinnerTeamIndex: request.winning_team_index,
+        SeriesScore: this.haloService.getSeriesScore(sortedSeries, locale),
+        Source: "neatqueue",
+        CreatedAt: nowEpoch,
+        UpdatedAt: nowEpoch,
+      };
+
+      const gamertagMap = await this.haloService.getPlayerXuidsToGametags(sortedSeries);
+      const allXuids = this.getSortedDistinctXuids(sortedSeries);
+      const associations = await this.databaseService.getDiscordAssociationsByXboxId(allXuids);
+      const xuidToDiscordId = new Map(associations.map((association) => [association.XboxId, association.DiscordId]));
+
+      const { gamesRows, gamePlayerRows, seriesPlayerRows } = await this.buildLeaderboardRows({
+        guildId: request.guild,
+        queueNumber: request.match_number,
+        queueChannelId: neatQueueConfig.ChannelId,
+        winnerTeamIndex: request.winning_team_index,
+        sortedSeries,
+        gamertagMap,
+        xuidToDiscordId,
+      });
+      await this.databaseService.upsertLeaderboardSeriesDataBatch({
+        series: seriesRow,
+        games: gamesRows,
+        gamePlayers: gamePlayerRows,
+        seriesPlayers: seriesPlayerRows,
+      });
+    } catch (error) {
+      this.logService.warn(
+        error,
+        new Map([
+          ["guildId", request.guild],
+          ["queueNumber", request.match_number.toString()],
+          ["reason", "Failed to persist leaderboard series data"],
+        ]),
+      );
+    }
+  }
+
+  private toEpochSeconds(value: string): number | null {
+    const parsed = Date.parse(value);
+    if (Number.isNaN(parsed)) {
+      return null;
+    }
+
+    return Math.floor(parsed / 1000);
+  }
+
+  private getAverageLifeSeconds(averageLifeDuration: string): number {
+    try {
+      return getDurationInSeconds(averageLifeDuration);
+    } catch {
+      return 0;
+    }
+  }
+
+  private getSortedDistinctXuids(series: MatchStats[]): string[] {
+    const xuids = new Set<string>();
+
+    for (const match of series) {
+      for (const player of match.Players) {
+        if (player.PlayerType !== 1) {
+          continue;
+        }
+
+        xuids.add(getPlayerXuid(player));
+      }
+    }
+
+    return [...xuids].sort((left, right) => left.localeCompare(right));
+  }
+
+  private async buildLeaderboardRows({
+    guildId,
+    queueNumber,
+    queueChannelId,
+    winnerTeamIndex,
+    sortedSeries,
+    gamertagMap,
+    xuidToDiscordId,
+  }: {
+    guildId: string;
+    queueNumber: number;
+    queueChannelId: string;
+    winnerTeamIndex: number;
+    sortedSeries: MatchStats[];
+    gamertagMap: Map<string, string>;
+    xuidToDiscordId: Map<string, string>;
+  }): Promise<{
+    gamesRows: LeaderboardGamesRow[];
+    gamePlayerRows: LeaderboardGamePlayersRow[];
+    seriesPlayerRows: LeaderboardSeriesPlayersRow[];
+  }> {
+    const nowEpoch = Math.floor(Date.now() / 1000);
+    const gamesRows: LeaderboardGamesRow[] = [];
+    const gamePlayerRows: LeaderboardGamePlayersRow[] = [];
+    const playersByXuid = new Map<string, LeaderboardSeriesPlayersRow>();
+    const participationByXuid = new Map<string, boolean[]>();
+    const gameTypeAndMapByMatchId = new Map<string, { gameType: string; gameMap: string }>();
+    const gameTypesAndMaps = await Promise.all(
+      sortedSeries.map(async (match) => ({
+        matchId: match.MatchId,
+        ...(await this.haloService.getGameTypeAndMapParts(match.MatchInfo)),
+      })),
+    );
+    for (const gameTypeAndMap of gameTypesAndMaps) {
+      gameTypeAndMapByMatchId.set(gameTypeAndMap.matchId, {
+        gameType: gameTypeAndMap.gameType,
+        gameMap: gameTypeAndMap.gameMap,
+      });
+    }
+
+    for (const [index, match] of sortedSeries.entries()) {
+      const { gameType, gameMap } = Preconditions.checkExists(
+        gameTypeAndMapByMatchId.get(match.MatchId),
+        "Expected resolved game type and map",
+      );
+      const startedAt = this.toEpochSeconds(match.MatchInfo.StartTime) ?? nowEpoch;
+      const endedAt = this.toEpochSeconds(match.MatchInfo.EndTime) ?? nowEpoch;
+      const team0Score = match.Teams.find((team) => team.TeamId === 0)?.Stats.CoreStats.Score ?? null;
+      const team1Score = match.Teams.find((team) => team.TeamId === 1)?.Stats.CoreStats.Score ?? null;
+
+      gamesRows.push({
+        MatchId: match.MatchId,
+        GuildId: guildId,
+        QueueNumber: queueNumber,
+        QueueChannelId: queueChannelId,
+        GameIndexInSeries: index,
+        GameVariantCategory: match.MatchInfo.GameVariantCategory,
+        ModeName: gameType,
+        MapName: gameMap,
+        MapAssetId: match.MatchInfo.MapVariant.AssetId,
+        MapVersionId: match.MatchInfo.MapVariant.VersionId,
+        Team0Score: team0Score,
+        Team1Score: team1Score,
+        StartedAt: startedAt,
+        EndedAt: endedAt,
+        CreatedAt: nowEpoch,
+      });
+
+      for (const player of match.Players) {
+        if (player.PlayerType !== 1) {
+          continue;
+        }
+
+        const xuid = getPlayerXuid(player);
+        const participation = participationByXuid.get(xuid);
+        if (participation == null) {
+          const initialParticipation = Array<boolean>(sortedSeries.length).fill(false);
+          initialParticipation[index] = true;
+          participationByXuid.set(xuid, initialParticipation);
+        } else {
+          participation[index] = true;
+        }
+
+        const teamStats = Preconditions.checkExists(
+          player.PlayerTeamStats.find((candidate) => match.Teams.some((team) => team.TeamId === candidate.TeamId)),
+          "Expected player team stats",
+        );
+        const coreStats = teamStats.Stats.CoreStats;
+        const deaths = coreStats.Deaths;
+
+        gamePlayerRows.push({
+          MatchId: match.MatchId,
+          GuildId: guildId,
+          QueueNumber: queueNumber,
+          QueueChannelId: queueChannelId,
+          XboxXuid: xuid,
+          DiscordUserId: xuidToDiscordId.get(xuid) ?? null,
+          GamertagSnapshot: gamertagMap.get(xuid) ?? "Unknown",
+          TeamId: teamStats.TeamId,
+          PresentAtBeginning: player.ParticipationInfo.PresentAtBeginning ? 1 : 0,
+          RankInMatch: player.Rank,
+          PersonalScore: coreStats.PersonalScore,
+          Kills: coreStats.Kills,
+          Deaths: deaths,
+          Assists: coreStats.Assists,
+          Kda: coreStats.KDA,
+          Accuracy: coreStats.Accuracy,
+          ShotsHit: coreStats.ShotsHit,
+          ShotsFired: coreStats.ShotsFired,
+          DamageDealt: coreStats.DamageDealt,
+          DamageTaken: coreStats.DamageTaken,
+          DamageRatio: getSafeRatioValue(coreStats.DamageDealt, coreStats.DamageTaken),
+          AvgLifeSeconds: this.getAverageLifeSeconds(coreStats.AverageLifeDuration),
+          AvgDamagePerLife: getSafeRatioValue(coreStats.DamageDealt, deaths),
+          ObjectiveStatsJson: JSON.stringify(teamStats.Stats),
+          MedalsJson: JSON.stringify(coreStats.Medals),
+          CreatedAt: nowEpoch,
+        });
+
+        const existing = playersByXuid.get(xuid);
+        if (existing == null) {
+          playersByXuid.set(xuid, {
+            GuildId: guildId,
+            QueueNumber: queueNumber,
+            QueueChannelId: queueChannelId,
+            XboxXuid: xuid,
+            DiscordUserId: xuidToDiscordId.get(xuid) ?? null,
+            GamertagSnapshot: gamertagMap.get(xuid) ?? "Unknown",
+            TeamId: teamStats.TeamId,
+            PresentAtBeginningCount: player.ParticipationInfo.PresentAtBeginning ? 1 : 0,
+            SubstituteInCount: player.ParticipationInfo.PresentAtBeginning ? 0 : 1,
+            SubstituteOutCount: 0,
+            GamesPlayedCount: 1,
+            SeriesWon: teamStats.TeamId === winnerTeamIndex ? 1 : 0,
+            CreatedAt: nowEpoch,
+          });
+          continue;
+        }
+
+        existing.GamesPlayedCount += 1;
+        if (player.ParticipationInfo.PresentAtBeginning) {
+          existing.PresentAtBeginningCount += 1;
+        } else {
+          existing.SubstituteInCount += 1;
+        }
+      }
+    }
+
+    for (const row of playersByXuid.values()) {
+      const participation = Preconditions.checkExists(
+        participationByXuid.get(row.XboxXuid),
+        "Expected player participation across series",
+      );
+      let substituteOutCount = 0;
+      for (let i = 0; i < participation.length - 1; i += 1) {
+        if (participation[i] === true && participation[i + 1] !== true) {
+          substituteOutCount += 1;
+        }
+      }
+      row.SubstituteOutCount = substituteOutCount;
+    }
+
+    const seriesPlayerRows = [...playersByXuid.values()].sort((left, right) =>
+      left.XboxXuid.localeCompare(right.XboxXuid),
+    );
+
+    return {
+      gamesRows,
+      gamePlayerRows,
+      seriesPlayerRows,
+    };
+  }
+}
